Changed data types in migratios because of cleardb

+++ database/migrations/2019_08_24_214655_create_emails_table.php
@@ -14,10 +14,10 @@ class CreateEmailsTable extends Migration
     public function up()
     {
         Schema::create('emails', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->text('subject');
             $table->longText('body');
-            $table->unsignedInteger('user_id');
+            $table->integer('user_id');
             $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
